Group the apiCredentials modules.
