+++ src/gateway/client.ts
