+++ tests/sharedOperations.test.ts
@@ -3,8 +3,8 @@ import { mkdtempSync, rmSync } from 'node:fs';
 import { join } from 'node:path';
 import { tmpdir } from 'node:os';
 import { EncryptedStorage } from '../src/encryptedStorage.js';
-import { ApiCredentialStore } from '../src/apiCredentialStore.js';
-import { ApiCredentialStatus } from '../src/apiCredentials.js';
+import { ApiCredentialStore } from '../src/apiCredentials/store.js';
+import { ApiCredentialStatus } from '../src/apiCredentials/base.js';
 import { SlackApiCredentials } from '../src/services/slack.js';
 import { NoCurlCredentialsNotSupportedError, Service } from '../src/services/core/base.js';
 import { RegisteredService } from '../src/services/core/registered.js';
